class page code restructure and bug fixing
