+++ content-scripts/src/pages/class_page.js
@@ -1,3 +1,5 @@
+import { fetchSigarraPage } from "../modules/utilities/pageUtils";
+
 const removeExtras = () => {
     document.querySelector("#conteudoinner > h2").remove();
     document.querySelector("#conteudoinner > table:nth-child(5)").remove();
@@ -12,11 +14,16 @@ const titleClick = (table) => {
     }
 }
 
-const createPhotosDialog = (table, index) => {
+const createPhotosDialog = async (title, index) => {
     const dialog = document.createElement("dialog");
     const wrapper = document.createElement("div");
     const closeButton = document.createElement("span");
-    
+
+    const url = getPhotosLink(title);
+
+    const html = await fetchSigarraPage(url);
+    const table = html.querySelector("#layout0 > table");
+
     dialog.id = "photosDialog" + index;
     closeButton.classList.add("closeDialog");
     closeButton.innerText = "X";
@@ -45,31 +52,11 @@ const createPhotosButton = (icon, dialog) => {
     return button;
 }
 
-const createClassPhotosButton = async (classLinks, title, classIndex, table) => {
-    const url = classLinks[1].href;
-    
-    const r = await fetch(url);
-
-    const decoder = new TextDecoder(
-        r.headers.get("Content-Type").replace("text/html; charset=", "")
-    );
-
-    const text = decoder.decode(await r.arrayBuffer());
-
-    const parser = new DOMParser();
-    let html = parser.parseFromString(text, "text/html");
-
-    const photosTable = html.querySelector("#layout0 > table");
-
-    const photosDialog = createPhotosDialog(photosTable, classIndex);
-
-    const photosButton = createPhotosButton(classLinks[1].children[0], photosDialog);
-    
-    table.appendChild(photosDialog);
-    title.appendChild(photosButton);
+const getPhotosLink = (title) => {
+    return title.children[1].href;
 }
 
-const editTitle = async (title, table, classIndex) => {   
+const editTitle = async (title, table, dialog) => {   
     const titleText = document.createElement("h3");
     const titleContent = title.children[0];
     const className = titleContent.innerText.replaceAll(String.fromCharCode(160), '');
@@ -81,13 +68,14 @@ const editTitle = async (title, table, classIndex) => {
     titleText.innerText = className;
     title.appendChild(classLinks[0]); // email button
     
-    await createClassPhotosButton(classLinks, title, classIndex, table); // photos modal
+    const photosButton = createPhotosButton(classLinks[1].children[0], dialog); // photos button
+    title.appendChild(photosButton);
 
     titleContent.remove();
-    title.addEventListener("click", () => titleClick(title, table));
+    title.addEventListener("click", () => titleClick(table));
 }
 
-const groupClasses = () => {
+const groupClasses = async () => {
     const parent = document.querySelector("#conteudoinner");
     let titleIndex = 5, tableIndex = 6, classIndex = 0;
     let title = document.querySelector(`#conteudoinner > h3:nth-child(${titleIndex})`);
@@ -97,7 +85,8 @@ const groupClasses = () => {
         const groupElement = document.createElement("section");
         const tableWrapperElement = document.createElement("div");
         const titleWrapperElement = document.createElement("div");
-        
+        const photosDialog = await createPhotosDialog(title, classIndex);
+
         parent.insertBefore(groupElement, title);
         groupElement.classList.add("classWrapper");
         tableWrapperElement.classList.add("tableWrapper");
@@ -109,8 +98,9 @@ const groupClasses = () => {
         tableWrapperElement.appendChild(table);
         groupElement.appendChild(titleWrapperElement);
         groupElement.appendChild(tableWrapperElement);
+        groupElement.appendChild(photosDialog);
         
-        editTitle(titleWrapperElement, tableWrapperElement, classIndex);
+        editTitle(titleWrapperElement, tableWrapperElement, photosDialog);
         tableWrapperElement.dataset.expand = "true";
 
         titleIndex++;
